Remove all nZR zombies
The nZR devs don't know how to code and everything they touch breaks. The only reason we used the moo base is because my admin Berb went ahead without my permission and added it. We will no longer be using anything nZR.
