--- gamemodes/nzombies/entities/entities/nz_zombie_special_base_moo.lua
@@ -1,6 +1,0 @@
--- Just gonna have to be a copy paste of nz_zombie_special_base since Gmod doesn't let us derive two different bases (at least I don't think it does..)
-AddCSLuaFile()
-
-ENT.Author = "Ethorbit"
-ENT.Base = "nz_zombiebase_moo"
-ENT.IgnoreDistractions = true -- Ignore stuff like monkey bombs, we almost always only care about the player.
