+++ gamemodes/nzombies/entities/entities/nz_zombie_special_burning.lua
@@ -25,100 +25,33 @@ nzConfig.AddValidEnemy("nz_zombie_special_burning", {
     end
 })
 
-ENT.Base = "nz_zombie_walker_derriese_supersprint"
+ENT.Base = "nz_zombie_walker"
 ENT.PrintName = "Burning Walker"
 ENT.Category = "Brainz"
-ENT.Author = "GhostlyMoo"
+ENT.Author = "Lolle"
 
 function ENT:StatsInitialize()
     if SERVER then
-        dying = false
-		if nzRound:GetNumber() == -1 then
-			self:SetRunSpeed( math.random(30, 300) )
+        if nzRound:GetNumber() == -1 then
+            self:SetRunSpeed( math.random(20, 260) )
 
-			local hp = math.random(100, 1500)
-			self:SetHealth(hp)
-			self:SetMaxHealth(hp)
-		else
-			local speeds = nzRound:GetZombieSpeeds()
-			if speeds and !table.IsEmpty(speeds) then
-				self:SetRunSpeed( nzMisc.WeightedRandom(speeds) + 35 )
-			else
-				self:SetRunSpeed( 100 )
-			end
-			local hp = nzRound:GetZombieHealth() or 75
-			self:SetHealth(hp)
-			self:SetMaxHealth(hp)
-		end
+            local hp = math.random(75, 1000)
+            self:SetHealth(hp)
+            self:SetMaxHealth(hp)
+        else
+            local speeds = nzRound:GetZombieSpeeds()
+            if speeds then
+                self:SetRunSpeed( nzMisc.WeightedRandom(speeds) - 20 ) -- A bit slower here
+            end
+
+            local hp = nzRound:GetZombieHealth() or 75 
+            self:SetHealth(hp)
+            self:SetMaxHealth(hp)
+        end
         self:Flames( true )
+
+        self:SetEmergeSequenceIndex(math.random(#self.EmergeSequences))
     end
-end
-
-function ENT:OnSpawn()
-	local nav = navmesh.GetNavArea(self:GetPos(), 50)
-	if IsValid(nav) and nav:HasAttributes(NAV_MESH_NO_JUMP) then
-		self:SolidMaskDuringEvent(MASK_PLAYERSOLID)
-		self:CollideWhenPossible()
-	else
-		local SpawnMatSound = {
-			[MAT_DIRT] = "nz_moo/zombies/spawn/dirt/pfx_zm_spawn_dirt_0"..math.random(0,1)..".mp3",
-			[MAT_SNOW] = "nz_moo/zombies/spawn/snow/pfx_zm_spawn_snow_0"..math.random(0,1)..".mp3",
-			[MAT_SLOSH] = "nz_moo/zombies/spawn/mud/pfx_zm_spawn_mud_00.mp3",
-			[0] = "nz_moo/zombies/spawn/default/pfx_zm_spawn_default_00.mp3",
-		}
-		SpawnMatSound[MAT_GRASS] = SpawnMatSound[MAT_DIRT]
-		SpawnMatSound[MAT_SAND] = SpawnMatSound[MAT_DIRT]
-
-		local norm = (self:GetPos()):GetNormalized()
-		local tr = util.QuickTrace(self:GetPos(), norm*10, self)
-
-		self:SolidMaskDuringEvent(MASK_PLAYERSOLID)
-
-		--ParticleEffect("bo3_zombie_spawn",self:GetPos()+Vector(0,0,1),self:GetAngles(),self)
-		--self:EmitSound("nz/zombies/spawn/zm_spawn_dirt"..math.random(1,2)..".wav",80,math.random(95,105))
-	
-		self:SetSpecialAnimation(true)
-		self:SetIsBusy(true)
-		local seq = self:SelectSpawnSequence()
-
-		if IsValid(nav) and nav:HasAttributes(NAV_MESH_OBSTACLE_TOP) then
-			seq = self.UndercroftSequences[math.random(#self.UndercroftSequences)]
-		elseif IsValid(nav) and nav:HasAttributes(NAV_MESH_DONT_HIDE) then
-			seq = "nz_moo_wall_emerge_quick"
-		else
-			if tr.Hit then
-				local finalsound = SpawnMatSound[tr.MatType] or SpawnMatSound[0]
-				self:EmitSound(finalsound)
-			end
-			ParticleEffect("bo3_zombie_spawn",self:GetPos()+Vector(0,0,1),self:GetAngles(),self)
-			self:EmitSound("nz/zombies/spawn/zm_spawn_dirt"..math.random(1,2)..".wav",80,math.random(95,105))
-		end
-		if seq then
-			if IsValid(nav) and (nav:HasAttributes(NAV_MESH_OBSTACLE_TOP) or nav:HasAttributes(NAV_MESH_DONT_HIDE)) then
-				self:PlaySequenceAndMove(seq, {gravity = false})
-			else
-				self:PlaySequenceAndMove(seq, {gravity = true})
-			end
-			self:SetSpecialAnimation(false)
-			self:SetIsBusy(false)
-			self:CollideWhenPossible()
-		end
-	end
-end
-
-function ENT:Sound()
-	if self.BehindSoundDistance > 0 -- We have enabled behind sounds
-		and IsValid(self.Target)
-		and self.Target:IsPlayer() -- We have a target and it's a player within distance
-		and self:GetRangeTo(self.Target) <= self.BehindSoundDistance
-		and (self.Target:GetPos() - self:GetPos()):GetNormalized():Dot(self.Target:GetAimVector()) >= 0 then -- If the direction towards the player is same 180 degree as the player's aim (away from the zombie)
-			self:PlaySound(self.BehindSounds[math.random(#self.BehindSounds)], SNDLVL_140) -- Play the behind sound, and a bit louder!
-	elseif self.PassiveSounds then
-		self:PlaySound(self.PassiveSounds[math.random(#self.PassiveSounds)],100, math.random(65, 75)) --Lower pitched Vox cuz it sounds cooler
-	else
-		-- We still delay by max sound delay even if there was no sound to play
-		self.NextSound = CurTime() + self.SoundDelayMax
-	end
 end
 
 function ENT:OnTargetInAttackRange()
@@ -127,14 +60,15 @@ function ENT:OnTargetInAttackRange()
     atkData.dmghigh = 30
     atkData.dmgforce = Vector( 0, 0, 0 )
     self:Attack( atkData )
-    if self:IsValidTarget( self:GetTarget() ) and self:TargetInRange( self.AttackRange + 10 ) then
-       self:Explode( math.random( 50, 100 ) )
-    end
+    self:TimedEvent( 0.45, function()
+        if self:IsValidTarget( self:GetTarget() ) and self:TargetInRange( self.AttackRange + 10 ) then
+            self:Explode( math.random( 50, 100 ) )
+        end
+    end)
 end
 
-function ENT:PerformDeath(dmgInfo)
+function ENT:OnZombieDeath(dmgInfo)
     self:Explode( math.random( 25, 50 ))
-	self:EmitSound(self.DeathSounds[ math.random( #self.DeathSounds ) ], 50, math.random(75, 130))
-	self:BecomeRagdoll(dmgInfo)
+    self:EmitSound(self.DeathSounds[ math.random( #self.DeathSounds ) ], 50, math.random(75, 130))
+    self:BecomeRagdoll(dmgInfo)
 end
-
