+++ gamemodes/nzombies/entities/entities/nz_zombie_walker.lua
@@ -1,6 +1,6 @@
 AddCSLuaFile()
 
-nzRound:AddZombieType("WaW", "nz_zombie_walker", {
+nzRound:AddZombieType("OG", "nz_zombie_walker", {
     -- Set to false to disable the spawning of this zombie
     Valid = true,
     -- Allow you to scale damage on a per-hitgroup basis
@@ -29,875 +29,294 @@ nzRound:AddZombieType("WaW", "nz_zombie_walker", {
             end
         end
     end
-}) 
+})
 
-ENT.Base = "nz_zombiebase_moo"
-ENT.Type = "nextbot"
+ENT.Base = "nz_zombiebase"
+ENT.PrintName = "Walker"
 ENT.Category = "Brainz"
-ENT.Author = "GhostlyMoo"
-ENT.Spawnable = true
+ENT.Author = "Lolle"
 
---if CLIENT then return end -- Client doesn't really need anything beyond the basics
+DEFINE_BASECLASS(ENT.Base)
 
-ENT.SpeedBasedSequences = true
-ENT.IsMooZombie = true
-ENT.RedEyes = true
+function ENT:SetupDataTables()
+    self:NetworkVar("Int", 0, "EmergeSequenceIndex")
+    BaseClass.SetupDataTables(self)
+    --self:NetworkVar("Bool", 1, "Decapitated")
+end
 
 ENT.Models = {
-	{Model = "models/moo/_codz_ports/t4/moo_codz_t4_honorguard.mdl", Skin = 0, Bodygroups = {0,0}},
-	{Model = "models/moo/_codz_ports/t4/moo_codz_t4_honorguard.mdl", Skin = 1, Bodygroups = {0,0}},
-	{Model = "models/moo/_codz_ports/t4/moo_codz_t4_honorguard.mdl", Skin = 2, Bodygroups = {0,0}},
-	{Model = "models/moo/_codz_ports/t4/moo_codz_t4_honorguard.mdl", Skin = 3, Bodygroups = {0,0}},
-
-	{Model = "models/moo/_codz_ports/t4/moo_codz_t4_honorguard.mdl", Skin = 4, Bodygroups = {0,0}},
-	{Model = "models/moo/_codz_ports/t4/moo_codz_t4_honorguard.mdl", Skin = 5, Bodygroups = {0,0}},
-	{Model = "models/moo/_codz_ports/t4/moo_codz_t4_honorguard.mdl", Skin = 6, Bodygroups = {0,0}},
-	{Model = "models/moo/_codz_ports/t4/moo_codz_t4_honorguard.mdl", Skin = 7, Bodygroups = {0,0}},
-	
-	{Model = "models/moo/_codz_ports/t4/moo_codz_t4_honorguard.mdl", Skin = 8, Bodygroups = {0,0}},
-}
-
-local spawnslow = {"nz_spawn_ground_v1", "nz_spawn_ground_ad_v2", "nz_spawn_ground_v2", "nz_spawn_ground_v2_altb"}
-local spawnrun = {"nz_spawn_ground_v1_run"}
-local spawnfast = {"nz_spawn_ground_climbout_fast"}
-local spawnsuperfast = {"nz_spawn_ground_quickrise_v1", "nz_spawn_ground_quickrise_v2", "nz_spawn_ground_quickrise_v3"}
-
-ENT.DeathSequences = {
-	"nz_death_1",
-	"nz_death_2",
-	"nz_death_3",
-	"nz_death_f_1",
-	"nz_death_f_2",
-	"nz_death_f_3",
-	"nz_death_f_4",
-	"nz_death_f_5",
-	"nz_death_f_6",
-	"nz_death_f_7",
-	"nz_death_f_8",
-	"nz_death_f_9",
-	"nz_death_f_10",
-	"nz_death_f_11",
-	"nz_death_f_12",
-	"nz_death_f_13",
-	"nz_death_fallback",
-}
-
-ENT.CrawlDeathSequences = {
-	"nz_crawl_death_v1",
-	"nz_crawl_death_v2",
-}
-
-ENT.ElectrocutionSequences = {
-	"nz_death_elec_1",
-	"nz_death_elec_2",
-	"nz_death_elec_3",
-	"nz_death_elec_4",
-}
-
-ENT.BarricadeTearSequences = {
-	"nz_legacy_door_tear_high",
-	"nz_legacy_door_tear_low",
-	"nz_legacy_door_tear_left",
-	"nz_legacy_door_tear_right",
-}
-
-local CrawlAttackSequences = {
-	{seq = "nz_crawl_attack_v1", dmgtimes = {0.75, 1.65}},
-	{seq = "nz_crawl_attack_v2", dmgtimes = {0.65}},
-}
-
-local CrawlJumpSequences = {
-	{seq = "nz_barricade_crawl_1"},
-	{seq = "nz_barricade_crawl_2"},
-}
-
-local SlowClimbUp36 = {
-	"nz_traverse_climbup36"
-}
-local SlowClimbUp48 = {
-	"nz_traverse_climbup48"
-}
-local SlowClimbUp72 = {
-	"nz_traverse_climbup72"
-}
-local SlowClimbUp96 = {
-	"nz_traverse_climbup96"
-}
-local SlowClimbUp128 = {
-	"nz_traverse_climbup128"
-}
-local SlowClimbUp160 = {
-	"nz_traverse_climbup160"
-}
-local FastClimbUp36 = {
-	"nz_traverse_fast_climbup36"
-}
-local FastClimbUp48 = {
-	"nz_traverse_fast_climbup48"
-}
-local FastClimbUp72 = {
-	"nz_traverse_fast_climbup72"
-}
-local FastClimbUp96 = {
-	"nz_traverse_fast_climbup96"
-}
-local ClimbUp200 = {
-	"nz_traverse_climbup200"
+    "models/nz_zombie/zombie_rerig_animated.mdl",
 }
 
 local AttackSequences = {
-	{seq = "nz_attack_stand_ad_1"},
-	{seq = "nz_attack_stand_au_1"},
-	{seq = "nz_legacy_attack_v3"},
-	{seq = "nz_legacy_attack_v6"},
-	{seq = "nz_legacy_attack_v4"},
-	{seq = "nz_legacy_attack_v11"},
-	{seq = "nz_legacy_attack_v12"},
-	{seq = "nz_legacy_attack_superwindmill"},
-	{seq = "nz_fwd_ad_attack_v1"},
-	{seq = "nz_fwd_ad_attack_v2"},
-	{seq = "nz_t8_attack_stand_larm_1"},
-	{seq = "nz_t8_attack_stand_larm_2"},
-	{seq = "nz_t8_attack_stand_larm_3"},
-	{seq = "nz_t8_attack_stand_rarm_1"},
-	{seq = "nz_t8_attack_stand_rarm_2"},
-	{seq = "nz_t8_attack_stand_rarm_3"},
+    {seq = "nz_stand_attack1", dmgtimes = {0.75, 1.25}},
+    {seq = "nz_stand_attack2", dmgtimes = {0.3}},
+    {seq = "nz_stand_attack3", dmgtimes = {0.8}},
+    {seq = "nz_stand_attack4", dmgtimes = {0.4, 0.8}},
 }
-
 local WalkAttackSequences = {
-	{seq = "nz_walk_ad_attack_v1"}, -- Quick single swipe
-	{seq = "nz_walk_ad_attack_v2"}, -- Slowish double swipe
-	{seq = "nz_walk_ad_attack_v3"}, -- Slowish single swipe
-	{seq = "nz_walk_ad_attack_v4"}, -- Quickish double swipe
-	{seq = "nz_t8_attack_walk_larm_1"},
-	{seq = "nz_t8_attack_walk_rarm_3"},
-	{seq = "nz_t8_attack_walk_larm_2"},
-	{seq = "nz_t8_attack_walk_rarm_6"},
+    {seq = "nz_walk_attack1", dmgtimes = {0.3}},
+    {seq = "nz_walk_attack2", dmgtimes = {0.4, 0.9}},
+    {seq = "nz_walk_attack3", dmgtimes = {0.5}},
+    {seq = "nz_walk_attack4", dmgtimes = {0.4, 0.75}},
 }
-
 local RunAttackSequences = {
-	{seq = "nz_t8_attack_run_larm_1"},
-	{seq = "nz_t8_attack_run_larm_2"},
-	{seq = "nz_t8_attack_run_larm_3"},
-	{seq = "nz_t8_attack_run_larm_4"},
-	{seq = "nz_t8_attack_run_rarm_1"},
-	{seq = "nz_t8_attack_run_rarm_2"},
-	{seq = "nz_t8_attack_run_rarm_3"},
-	{seq = "nz_t8_attack_run_rarm_4"},
+    {seq = "nz_run_attack1", dmgtimes = {0.3}},
+    {seq = "nz_run_attack2", dmgtimes = {0.3, 0.65}},
+    {seq = "nz_run_attack3", dmgtimes = {0.3, 0.7}},
+    {seq = "nz_run_attack4", dmgtimes = {0.3, 0.8}},
+}
+local AttackSounds = {
+    "nz/zombies/attack/attack_00.wav",
+    "nz/zombies/attack/attack_01.wav",
+    "nz/zombies/attack/attack_02.wav",
+    "nz/zombies/attack/attack_03.wav",
+    "nz/zombies/attack/attack_04.wav",
+    "nz/zombies/attack/attack_05.wav",
+    "nz/zombies/attack/attack_06.wav",
+    "nz/zombies/attack/attack_07.wav",
+    "nz/zombies/attack/attack_08.wav",
+    "nz/zombies/attack/attack_09.wav",
+    "nz/zombies/attack/attack_10.wav",
+    "nz/zombies/attack/attack_11.wav",
+    "nz/zombies/attack/attack_12.wav",
+    "nz/zombies/attack/attack_13.wav",
+    "nz/zombies/attack/attack_14.wav",
+    "nz/zombies/attack/attack_15.wav",
+    "nz/zombies/attack/attack_16.wav",
+    "nz/zombies/attack/attack_17.wav",
+    "nz/zombies/attack/attack_18.wav",
+    "nz/zombies/attack/attack_19.wav",
+    "nz/zombies/attack/attack_20.wav",
+    "nz/zombies/attack/attack_21.wav",
+    "nz/zombies/attack/attack_22.wav",
 }
 
-local StinkyRunAttackSequences = {
-	{seq = "nz_run_ad_attack_v1"},
-	{seq = "nz_run_ad_attack_v2"},
-	{seq = "nz_run_ad_attack_v3"},
-	{seq = "nz_run_ad_attack_v4"},
+local WalkSounds = {
+    "nz/zombies/ambient/ambient_00.wav",
+    "nz/zombies/ambient/ambient_01.wav",
+    "nz/zombies/ambient/ambient_02.wav",
+    "nz/zombies/ambient/ambient_03.wav",
+    "nz/zombies/ambient/ambient_04.wav",
+    "nz/zombies/ambient/ambient_05.wav",
+    "nz/zombies/ambient/ambient_06.wav",
+    "nz/zombies/ambient/ambient_07.wav",
+    "nz/zombies/ambient/ambient_08.wav",
+    "nz/zombies/ambient/ambient_09.wav",
+    "nz/zombies/ambient/ambient_10.wav",
+    "nz/zombies/ambient/ambient_11.wav",
+    "nz/zombies/ambient/ambient_12.wav",
+    "nz/zombies/ambient/ambient_13.wav",
+    "nz/zombies/ambient/ambient_14.wav",
+    "nz/zombies/ambient/ambient_15.wav",
+    "nz/zombies/ambient/ambient_16.wav",
+    "nz/zombies/ambient/ambient_17.wav",
+    "nz/zombies/ambient/ambient_18.wav",
+    "nz/zombies/ambient/ambient_19.wav",
+    "nz/zombies/ambient/ambient_20.wav"
 }
 
-local SprintAttackSequences = {
-	{seq = "nz_t8_attack_sprint_larm_1"},
-	{seq = "nz_t8_attack_sprint_larm_2"},
-	{seq = "nz_t8_attack_sprint_larm_3"},
-	{seq = "nz_t8_attack_sprint_larm_4"},
-	{seq = "nz_t8_attack_sprint_rarm_1"},
-	{seq = "nz_t8_attack_sprint_rarm_2"},
-	{seq = "nz_t8_attack_sprint_rarm_3"},
-	{seq = "nz_t8_attack_sprint_rarm_4"},
-}
-
-local SuperSprintAttackSequences = {
-	{seq = "nz_t8_attack_supersprint_larm_1"},
-	{seq = "nz_t8_attack_supersprint_larm_2"},
-	{seq = "nz_t8_attack_supersprint_rarm_1"},
-	{seq = "nz_t8_attack_supersprint_rarm_2"},
+local RunSounds = {
+    "nz/zombies/sprint2/sprint0.wav",
+    "nz/zombies/sprint2/sprint1.wav",
+    "nz/zombies/sprint2/sprint2.wav",
+    "nz/zombies/sprint2/sprint3.wav",
+    "nz/zombies/sprint2/sprint4.wav",
+    "nz/zombies/sprint2/sprint5.wav",
+    "nz/zombies/sprint2/sprint6.wav",
+    "nz/zombies/sprint2/sprint7.wav",
+    "nz/zombies/sprint2/sprint8.wav"
 }
 
 local JumpSequences = {
-	{seq = "nz_barricade_trav_walk_1"},
-	{seq = "nz_barricade_trav_walk_2"},
-	{seq = "nz_barricade_trav_walk_3"},
+    {seq = "nz_barricade1", speed = 15, time = 2.7},
+    {seq = "nz_barricade2", speed = 15, time = 2.4},
+    {seq = "nz_barricade_fast1", speed = 15, time = 1.8},
+    {seq = "nz_barricade_fast2", speed = 35, time = 4},
 }
-
-local RunJumpSequences = {
-	{seq = "nz_barricade_run_1"},
-}
-
 local SprintJumpSequences = {
-	{seq = "nz_barricade_sprint_1"},
-	{seq = "nz_barricade_sprint_2"},
-}
-local walksounds = {
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_00.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_01.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_02.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_03.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_04.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_05.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_06.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_07.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_08.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_09.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_10.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_11.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_12.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_13.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_14.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_15.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_16.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_17.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_18.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_19.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/amb/amb_20.mp3"),
-
+    {seq = "nz_barricade_sprint1", speed = 50, time = 1.9},
+    {seq = "nz_barricade_sprint2", speed = 35, time = 1.9},
 }
 
-local runsounds = {
-	Sound("nz_moo/zombies/vox/_classic/sprint/sprint_00.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/sprint/sprint_01.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/sprint/sprint_02.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/sprint/sprint_03.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/sprint/sprint_04.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/sprint/sprint_05.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/sprint/sprint_06.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/sprint/sprint_07.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/sprint/sprint_08.mp3"),
-	
-	Sound("nz_moo/zombies/vox/_proto/sprint/sprint_04.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/sprint/sprint_05.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/sprint/sprint_06.mp3"),
-	Sound("nz_moo/zombies/vox/_proto/sprint/sprint_07.mp3"),
+ENT.ActStages = {
+    [1] = {
+        act = ACT_WALK,
+        minspeed = 5,
+        attackanims = WalkAttackSequences,
+        -- no attackhitsounds, just use ENT.AttackHitSounds for all act stages
+        sounds = WalkSounds,
+        barricadejumps = JumpSequences,
+    },
+    [2] = {
+        act = ACT_WALK_ANGRY,
+        minspeed = 40,
+        attackanims = WalkAttackSequences,
+        sounds = WalkSounds,
+        barricadejumps = JumpSequences,
+    },
+    [3] = {
+        act = ACT_RUN,
+        minspeed = 100,
+        attackanims = RunAttackSequences,
+        sounds = RunSounds,
+        barricadejumps = SprintJumpSequences,
+    },
+    [4] = {
+        act = ACT_SPRINT,
+        minspeed = 160,
+        attackanims = RunAttackSequences,
+        sounds = RunSounds,
+        barricadejumps = SprintJumpSequences,
+    },
 }
 
--- This is a very large and messy looking table... But it gets the job done.
-ENT.SequenceTables = {
-	{Threshold = 0, Sequences = {
-		{
-			SpawnSequence = {spawnslow},
-			MovementSequence = {
-				"nz_walk_ad1",
-				"nz_walk_ad2",
-				"nz_walk_ad3",
-				"nz_walk_ad4",
-				"nz_walk_ad7",
-				"nz_walk_ad5",
-				"nz_walk_ad6",
-				"nz_walk_ad19",
-				"nz_walk_ad20",
-				"nz_walk_ad21",
-				"nz_walk_ad22",
-				"nz_walk_ad23",
-				"nz_walk_ad24",
-				"nz_walk_ad25",
-				"nz_walk_au_goose",
-				"nz_legacy_walk_dazed",
-				--"nz_legacy_jap_walk_v1",
-				--"nz_legacy_jap_walk_v2",
-				--"nz_legacy_jap_walk_v3",
-				--"nz_legacy_jap_walk_v4",
-				"nz_legacy_walk_v9",
-			},
-			LowgMovementSequence = {
-				"nz_walk_lowg_v1",
-				"nz_walk_lowg_v2",
-				"nz_walk_lowg_v3",
-				"nz_walk_lowg_v4",
-			},
-			BlackholeMovementSequence = {
-				"nz_blackhole_1",
-				"nz_blackhole_2",
-				"nz_blackhole_3",
-			},
-			CrawlMovementSequence = {
-				"nz_crawl_slow_v1",
-				"nz_crawl_slow_v2",
-				"nz_crawl_slow_v3",
-				"nz_crawl_v1",
-				"nz_crawl_v2",
-				"nz_crawl_v5",
-				"nz_crawl_sprint_v1",
-				"nz_crawl_on_hands",
-				"nz_crawl_on_hands_c",
-			},
-			FireMovementSequence = {
-				"nz_firestaff_walk_v1",
-				"nz_firestaff_walk_v2",
-				"nz_firestaff_walk_v3",
-			},
-			TurnedMovementSequence = {
-				--"nz_pb_zombie_sprint_v6",
-				--"nz_pb_zombie_sprint_v7",
-				"nz_pb_zombie_sprint_v8", -- The Tranzit Sprinter one.
-				--"nz_pb_zombie_sprint_v9",
-			},
-			AttackSequences = {WalkAttackSequences},
-			StandAttackSequences = {AttackSequences},
-			CrawlAttackSequences = {CrawlAttackSequences},
+ENT.RedEyes = true
 
-			JumpSequences = {JumpSequences},
-			CrawlJumpSequences = {CrawlJumpSequences},
-
-			Climb36 = {SlowClimbUp36},
-			Climb48 = {SlowClimbUp48},
-			Climb72 = {SlowClimbUp72},
-			Climb96 = {SlowClimbUp96},
-			Climb120 = {SlowClimbUp128},
-			Climb160 = {SlowClimbUp160},
-			Climb200 = {ClimbUp200},
-
-			PassiveSounds = {walksounds},
-		},
-		{
-			SpawnSequence = {spawnslow},
-			MovementSequence = {
-				"nz_walk_au1",
-				"nz_walk_au2",
-				"nz_walk_au3",
-				"nz_walk_au4",
-				"nz_walk_au5",
-				"nz_walk_au6",
-				"nz_walk_au7",
-				"nz_walk_au8",
-				"nz_walk_au10",
-				"nz_walk_au11",
-				"nz_walk_au12",
-				"nz_walk_au13",
-				"nz_walk_au15",
-				"nz_walk_au20",
-				"nz_walk_au21",
-				"nz_walk_au23",
-				"nz_walk_au_goose", -- This is the goosestep walk aka marching anim that german soldier zombies use.
-				"nz_legacy_walk_dazed",
-				--"nz_legacy_jap_walk_v1",
-				--"nz_legacy_jap_walk_v2",
-				--"nz_legacy_jap_walk_v3",
-				--"nz_legacy_jap_walk_v4",
-				"nz_legacy_walk_v9",
-			},
-			LowgMovementSequence = {
-				"nz_walk_lowg_v1",
-				"nz_walk_lowg_v2",
-				"nz_walk_lowg_v3",
-				"nz_walk_lowg_v4",
-			},
-			BlackholeMovementSequence = {
-				"nz_blackhole_1",
-				"nz_blackhole_2",
-				"nz_blackhole_3",
-			},
-			CrawlMovementSequence = {
-				"nz_crawl_slow_v1",
-				"nz_crawl_slow_v2",
-				"nz_crawl_slow_v3",
-				"nz_crawl_v1",
-				"nz_crawl_v2",
-				"nz_crawl_v5",
-				"nz_crawl_sprint_v1",
-				"nz_crawl_on_hands",
-				"nz_crawl_on_hands_c",
-			},
-			FireMovementSequence = {
-				"nz_firestaff_walk_v1",
-				"nz_firestaff_walk_v2",
-				"nz_firestaff_walk_v3",
-			},
-			TurnedMovementSequence = {
-				--"nz_pb_zombie_sprint_v6",
-				--"nz_pb_zombie_sprint_v7",
-				"nz_pb_zombie_sprint_v8", -- The Tranzit Sprinter one.
-				--"nz_pb_zombie_sprint_v9",
-			},
-			AttackSequences = {WalkAttackSequences},
-			StandAttackSequences = {AttackSequences},
-			CrawlAttackSequences = {CrawlAttackSequences},
-
-			JumpSequences = {JumpSequences},
-			CrawlJumpSequences = {CrawlJumpSequences},
-
-			Climb36 = {SlowClimbUp36},
-			Climb48 = {SlowClimbUp48},
-			Climb72 = {SlowClimbUp72},
-			Climb96 = {SlowClimbUp96},
-			Climb120 = {SlowClimbUp128},
-			Climb160 = {SlowClimbUp160},
-			Climb200 = {ClimbUp200},
-
-			PassiveSounds = {walksounds},
-		}
-	}},
-	{Threshold = 36, Sequences = {
-		{
-			SpawnSequence = {spawnrun},
-			MovementSequence = {
-				"nz_walk_fast_ad1",
-				"nz_walk_fast_ad2",
-				"nz_walk_fast_ad3",
-				"nz_legacy_run_v1",
-				"nz_legacy_run_v3",
-				--"nz_legacy_jap_run_v1",
-				--"nz_legacy_jap_run_v2",
-				--"nz_legacy_jap_run_v4",
-				--"nz_legacy_jap_run_v5",
-				--"nz_legacy_jap_run_v6",
-				"nz_run_ad1",
-				"nz_run_ad2",
-				"nz_run_ad3",
-				"nz_run_ad4",
-				"nz_run_ad7",
-				"nz_run_ad8",
-				"nz_run_ad11",
-				"nz_run_ad12",
-				"nz_run_ad14",
-				"nz_run_ad20",
-				"nz_run_ad21",
-				"nz_run_ad22",
-				"nz_run_ad23",
-				"nz_run_ad24",
-			},
-			LowgMovementSequence = {
-				"nz_run_lowg_v1",
-				"nz_run_lowg_v2",
-				"nz_run_lowg_v3",
-				"nz_run_lowg_v4",
-			},
-			BlackholeMovementSequence = {
-				"nz_blackhole_1",
-				"nz_blackhole_2",
-				"nz_blackhole_3",
-			},
-			CrawlMovementSequence = {
-				"nz_crawl_slow_v1",
-				"nz_crawl_slow_v2",
-				"nz_crawl_slow_v3",
-				"nz_crawl_v1",
-				"nz_crawl_v2",
-				"nz_crawl_v5",
-				"nz_crawl_sprint_v1",
-				"nz_crawl_on_hands",
-				"nz_crawl_on_hands_c",
-			},
-			FireMovementSequence = {
-				"nz_firestaff_walk_v1",
-				"nz_firestaff_walk_v2",
-				"nz_firestaff_walk_v3",
-			},
-			TurnedMovementSequence = {
-				--"nz_pb_zombie_sprint_v6",
-				--"nz_pb_zombie_sprint_v7",
-				"nz_pb_zombie_sprint_v8", -- The Tranzit Sprinter one.
-				--"nz_pb_zombie_sprint_v9",
-			},
-			AttackSequences = {RunAttackSequences},
-			StandAttackSequences = {AttackSequences},
-			Bo3AttackSequences = {StinkyRunAttackSequences},
-			CrawlAttackSequences = {CrawlAttackSequences},
-
-			JumpSequences = {RunJumpSequences},
-			CrawlJumpSequences = {CrawlJumpSequences},
-
-			Climb36 = {SlowClimbUp36},
-			Climb48 = {SlowClimbUp48},
-			Climb72 = {SlowClimbUp72},
-			Climb96 = {SlowClimbUp96},
-			Climb120 = {SlowClimbUp128},
-			Climb160 = {SlowClimbUp160},
-			Climb200 = {ClimbUp200},
-
-			PassiveSounds = {runsounds},
-		},
-		{
-			SpawnSequence = {spawnrun},
-			MovementSequence = {
-				"nz_walk_fast_au1",
-				"nz_walk_fast_au2",
-				"nz_walk_fast_au3",
-				"nz_legacy_run_v1",
-				"nz_legacy_run_v3",
-				--"nz_legacy_jap_run_v1",
-				--"nz_legacy_jap_run_v2",
-				--"nz_legacy_jap_run_v4",
-				--"nz_legacy_jap_run_v5",
-				--"nz_legacy_jap_run_v6",
-				"nz_run_au1",
-				"nz_run_au2",
-				"nz_run_au3",
-				"nz_run_au4",
-				"nz_run_au5",
-				"nz_run_au9",
-				"nz_run_au11",
-				"nz_run_au13",
-				"nz_run_au20",
-				"nz_run_au21",
-				"nz_run_au22",
-				"nz_run_au23",
-				"nz_run_au24",
-			},
-			LowgMovementSequence = {
-				"nz_run_lowg_v1",
-				"nz_run_lowg_v2",
-				"nz_run_lowg_v3",
-				"nz_run_lowg_v4",
-			},
-			BlackholeMovementSequence = {
-				"nz_blackhole_1",
-				"nz_blackhole_2",
-				"nz_blackhole_3",
-			},
-			CrawlMovementSequence = {
-				"nz_crawl_slow_v1",
-				"nz_crawl_slow_v2",
-				"nz_crawl_slow_v3",
-				"nz_crawl_v1",
-				"nz_crawl_v2",
-				"nz_crawl_v5",
-				"nz_crawl_sprint_v1",
-				"nz_crawl_on_hands",
-				"nz_crawl_on_hands_c",
-			},
-			FireMovementSequence = {
-				"nz_firestaff_walk_v1",
-				"nz_firestaff_walk_v2",
-				"nz_firestaff_walk_v3",
-			},
-			TurnedMovementSequence = {
-				--"nz_pb_zombie_sprint_v6",
-				--"nz_pb_zombie_sprint_v7",
-				"nz_pb_zombie_sprint_v8", -- The Tranzit Sprinter one.
-				--"nz_pb_zombie_sprint_v9",
-			},
-			AttackSequences = {RunAttackSequences},
-			StandAttackSequences = {AttackSequences},
-			Bo3AttackSequences = {StinkyRunAttackSequences},
-			CrawlAttackSequences = {CrawlAttackSequences},
-
-			JumpSequences = {RunJumpSequences},
-			CrawlJumpSequences = {CrawlJumpSequences},
-
-			Climb36 = {SlowClimbUp36},
-			Climb48 = {SlowClimbUp48},
-			Climb72 = {SlowClimbUp72},
-			Climb96 = {SlowClimbUp96},
-			Climb120 = {SlowClimbUp128},
-			Climb160 = {SlowClimbUp160},
-			Climb200 = {ClimbUp200},
-
-			PassiveSounds = {runsounds},
-		}
-	}},
-	{Threshold = 71, Sequences = {
-		{
-			SpawnSequence = {spawnfast},
-			MovementSequence = {
-				"nz_legacy_sprint_v4",
-				"nz_legacy_sprint_v5",
-				--"nz_legacy_jap_run_v3",
-				"nz_sprint_ad1",
-				"nz_sprint_ad2",
-				"nz_sprint_ad3",
-				--"nz_sprint_ad4",
-				"nz_sprint_ad5",
-				"nz_sprint_ad21",
-				"nz_sprint_ad22",
-				"nz_sprint_ad23",
-				"nz_sprint_ad24",
-				"nz_fast_sprint_v1",
-				"nz_fast_sprint_v2",
-			},
-			LowgMovementSequence = {
-				"nz_sprint_lowg_v1",
-				"nz_sprint_lowg_v2",
-				"nz_sprint_lowg_v3",
-				"nz_sprint_lowg_v4",
-			},
-			BlackholeMovementSequence = {
-				"nz_blackhole_1",
-				"nz_blackhole_2",
-				"nz_blackhole_3",
-			},
-			CrawlMovementSequence = {
-				"nz_crawl_slow_v1",
-				"nz_crawl_slow_v2",
-				"nz_crawl_slow_v3",
-				"nz_crawl_v1",
-				"nz_crawl_v2",
-				"nz_crawl_v5",
-				"nz_crawl_sprint_v1",
-				"nz_crawl_on_hands",
-				"nz_crawl_on_hands_c",
-			},
-			FireMovementSequence = {
-				"nz_firestaff_walk_v1",
-				"nz_firestaff_walk_v2",
-				"nz_firestaff_walk_v3",
-			},
-			TurnedMovementSequence = {
-				--"nz_pb_zombie_sprint_v6",
-				--"nz_pb_zombie_sprint_v7",
-				"nz_pb_zombie_sprint_v8", -- The Tranzit Sprinter one.
-				--"nz_pb_zombie_sprint_v9",
-			},
-			AttackSequences = {SprintAttackSequences},
-			StandAttackSequences = {AttackSequences},
-			Bo3AttackSequences = {StinkyRunAttackSequences},
-			CrawlAttackSequences = {CrawlAttackSequences},
-
-			JumpSequences = {SprintJumpSequences},
-			CrawlJumpSequences = {CrawlJumpSequences},
-
-			Climb36 = {FastClimbUp36},
-			Climb48 = {FastClimbUp48},
-			Climb72 = {FastClimbUp72},
-			Climb96 = {FastClimbUp96},
-			Climb120 = {SlowClimbUp128},
-			Climb160 = {SlowClimbUp160},
-			Climb200 = {ClimbUp200},
-
-			PassiveSounds = {runsounds},
-		},
-		{
-			SpawnSequence = {spawnfast},
-			MovementSequence = {
-				"nz_legacy_sprint_v4",
-				"nz_legacy_sprint_v5",
-				--"nz_legacy_jap_run_v3",
-				"nz_sprint_au1",
-				"nz_sprint_au2",
-				"nz_sprint_au3",
-				--"nz_sprint_au4",
-				"nz_sprint_au20",
-				"nz_sprint_au21",
-				"nz_sprint_au22",
-				"nz_sprint_au25",
-				"nz_fast_sprint_v1",
-				"nz_fast_sprint_v2",
-			},
-			LowgMovementSequence = {
-				"nz_sprint_lowg_v1",
-				"nz_sprint_lowg_v2",
-				"nz_sprint_lowg_v3",
-				"nz_sprint_lowg_v4",
-			},
-			BlackholeMovementSequence = {
-				"nz_blackhole_1",
-				"nz_blackhole_2",
-				"nz_blackhole_3",
-			},
-			CrawlMovementSequence = {
-				"nz_crawl_slow_v1",
-				"nz_crawl_slow_v2",
-				"nz_crawl_slow_v3",
-				"nz_crawl_v1",
-				"nz_crawl_v2",
-				"nz_crawl_v5",
-				"nz_crawl_sprint_v1",
-				"nz_crawl_on_hands",
-				"nz_crawl_on_hands_c",
-			},
-			FireMovementSequence = {
-				"nz_firestaff_walk_v1",
-				"nz_firestaff_walk_v2",
-				"nz_firestaff_walk_v3",
-			},
-			TurnedMovementSequence = {
-				--"nz_pb_zombie_sprint_v6",
-				--"nz_pb_zombie_sprint_v7",
-				"nz_pb_zombie_sprint_v8", -- The Tranzit Sprinter one.
-				--"nz_pb_zombie_sprint_v9",
-			},
-			AttackSequences = {SprintAttackSequences},
-			StandAttackSequences = {AttackSequences},
-			Bo3AttackSequences = {StinkyRunAttackSequences},
-			CrawlAttackSequences = {CrawlAttackSequences},
-
-			JumpSequences = {SprintJumpSequences},
-			CrawlJumpSequences = {CrawlJumpSequences},
-
-			Climb36 = {FastClimbUp36},
-			Climb48 = {FastClimbUp48},
-			Climb72 = {FastClimbUp72},
-			Climb96 = {FastClimbUp96},
-			Climb120 = {SlowClimbUp128},
-			Climb160 = {SlowClimbUp160},
-			Climb200 = {ClimbUp200},
-
-			PassiveSounds = {runsounds},
-		}
-	}}
+ENT.ElectrocutionSequences = {
+    "nz_electrocuted1",
+    "nz_electrocuted2",
+    "nz_electrocuted3",
+    "nz_electrocuted4",
+    "nz_electrocuted5",
 }
-
-ENT.TauntSequences = {
-	"nz_taunt_v1",
-	"nz_taunt_v2",
-	"nz_taunt_v3",
-	"nz_taunt_v4",
-	"nz_taunt_v5",
-	"nz_taunt_v6",
-	"nz_taunt_v7",
-	"nz_taunt_v8",
-	"nz_taunt_v9"
+ENT.EmergeSequences = {
+    "nz_emerge1",
+    "nz_emerge2",
+    "nz_emerge3",
+    "nz_emerge4",
+    "nz_emerge5",
 }
-ENT.IdleSequence = "nz_idle_ad"
-
+ENT.AttackHitSounds = {
+    "nzr/zombies/attack/player_hit_0.wav",
+    "nzr/zombies/attack/player_hit_1.wav",
+    "nzr/zombies/attack/player_hit_2.wav",
+    "nzr/zombies/attack/player_hit_3.wav",
+    "nzr/zombies/attack/player_hit_4.wav",
+    "nzr/zombies/attack/player_hit_5.wav"
+}
+ENT.PainSounds = {
+    "nzr/zombies/death/nz_flesh_impact_0.wav",
+    "nzr/zombies/death/nz_flesh_impact_1.wav",
+    "nzr/zombies/death/nz_flesh_impact_2.wav",
+    "nzr/zombies/death/nz_flesh_impact_3.wav",
+    "nzr/zombies/death/nz_flesh_impact_4.wav"
+}
 ENT.DeathSounds = {
-	"nz_moo/zombies/vox/_classic/death/death_00.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_01.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_02.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_03.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_04.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_05.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_06.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_07.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_08.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_09.mp3",
-	"nz_moo/zombies/vox/_classic/death/death_10.mp3"
-}
-
-ENT.ElecSounds = {
-	"nz_moo/zombies/vox/_classic/elec/elec_00.mp3",
-	"nz_moo/zombies/vox/_classic/elec/elec_01.mp3",
-	"nz_moo/zombies/vox/_classic/elec/elec_02.mp3",
-	"nz_moo/zombies/vox/_classic/elec/elec_03.mp3",
-	"nz_moo/zombies/vox/_classic/elec/elec_04.mp3",
-	"nz_moo/zombies/vox/_classic/elec/elec_05.mp3"
-}
-
-ENT.NukeDeathSounds = {
-	"nz_moo/zombies/vox/nuke_death/soul_00.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_01.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_02.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_03.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_04.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_05.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_06.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_07.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_08.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_09.mp3",
-	"nz_moo/zombies/vox/nuke_death/soul_10.mp3"
-}
-
-ENT.AttackSounds = {
-	"nz_moo/zombies/vox/_proto/attack/attack_00.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_01.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_02.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_03.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_04.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_05.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_06.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_07.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_08.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_09.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_10.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_11.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_12.mp3",
-	"nz_moo/zombies/vox/_proto/attack/attack_13.mp3",
-
-	"nz_moo/zombies/vox/_classic/attack/attack_00.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_01.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_02.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_03.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_04.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_05.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_06.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_07.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_08.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_09.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_10.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_11.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_12.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_13.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_14.mp3",
-	"nz_moo/zombies/vox/_classic/attack/attack_15.mp3"
-}
-
-ENT.BehindSoundDistance = 200 -- When the zombie is within 200 units of a player, play these sounds instead
-ENT.BehindSounds = {
-	Sound("nz_moo/zombies/vox/_classic/behind/behind_00.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/behind/behind_01.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/behind/behind_02.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/behind/behind_03.mp3"),
-	Sound("nz_moo/zombies/vox/_classic/behind/behind_04.mp3"),
+    "nz/zombies/death/death_00.wav",
+    "nz/zombies/death/death_01.wav",
+    "nz/zombies/death/death_02.wav",
+    "nz/zombies/death/death_03.wav",
+    "nz/zombies/death/death_04.wav",
+    "nz/zombies/death/death_05.wav",
+    "nz/zombies/death/death_06.wav",
+    "nz/zombies/death/death_07.wav",
+    "nz/zombies/death/death_08.wav",
+    "nz/zombies/death/death_09.wav",
+    "nz/zombies/death/death_10.wav"
 }
 
 function ENT:StatsInitialize()
-	if SERVER then
-		if nzRound:GetNumber() == -1 then
-			self:SetRunSpeed( math.random(30, 300) )
+    if SERVER then
+        if nzRound:GetNumber() == -1 then
+            self:SetRunSpeed( math.random(30, 300) )
 
-			local hp = math.random(100, 1500)
-			self:SetHealth(hp)
-			self:SetMaxHealth(hp)
-		else
-			local speeds = nzRound:GetZombieSpeeds()
+            local hp = math.random(100, 1500)
+            self:SetHealth(hp)
+            self:SetMaxHealth(hp)
+        else
+            local speeds = nzRound:GetZombieSpeeds()
             if speeds and !table.IsEmpty(speeds) then
-				self:SetRunSpeed( nzMisc.WeightedRandom(speeds) )
-			else
-				self:SetRunSpeed( 100 )
-			end
-			local hp = nzRound:GetZombieHealth() or 75
-			self:SetHealth(hp)
-			self:SetMaxHealth(hp)
-		end
-	end
+                self:SetRunSpeed( nzMisc.WeightedRandom(speeds) )
+            else
+                self:SetRunSpeed( 100 )
+            end
+
+            local hp = nzRound:GetZombieHealth() or 75
+            self:SetHealth(hp)
+            self:SetMaxHealth(hp)
+        end
+
+        --Preselect the emerge sequnces for clientside use
+        self:SetEmergeSequenceIndex(math.random(#self.EmergeSequences))
+    end
 end
 
 function ENT:SpecialInit()
-	if CLIENT then
-	end
+    -- if SERVER then
+    --  self:TimedEvent(0.1, function()
+    --      self:SetNoDraw(true)
+
+    --      self:TimedEvent(0.15, function()
+    --          self:SetNoDraw(false)
+    --      end)
+    --  end)
+    -- end
+
+    if CLIENT then
+        --make them invisible for a really short duration to blend the emerge sequences
+        self:TimedEvent(0.1, function() -- Tiny delay just to make sure they are fully initialized
+            if string.find(self:GetSequenceName(self:GetSequence()), "nz_emerge") then
+                self:SetNoDraw(true)
+                self:TimedEvent( 0.15, function()
+                    self:SetNoDraw(false)
+                end)
+
+                self:SetRenderClipPlaneEnabled( true )
+                self:SetRenderClipPlane(self:GetUp(), self:GetUp():Dot(self:GetPos()))
+
+                --local _, dur = self:LookupSequence(self.EmergeSequences[self:GetEmergeSequenceIndex()])
+                local _, dur = self:LookupSequence(self.EmergeSequences[self:GetEmergeSequenceIndex()])
+                dur = dur - (dur * self:GetCycle()) -- Subtract the time we are already thruogh the animation
+                -- The above is important if the zombie only appears in PVS mid-way through the animation
+
+                self:TimedEvent( dur, function()
+                    self:SetRenderClipPlaneEnabled(false)
+                end)
+            end
+        end)
+
+    end
 end
 
 function ENT:OnSpawn()
-	local nav = navmesh.GetNavArea(self:GetPos(), 50)
-	if IsValid(nav) and nav:HasAttributes(NAV_MESH_NO_JUMP) then
-		self:SolidMaskDuringEvent(MASK_PLAYERSOLID)
-		self:CollideWhenPossible()
-	else
-		local SpawnMatSound = {
-			[MAT_DIRT] = "nz_moo/zombies/spawn/dirt/pfx_zm_spawn_dirt_0"..math.random(0,1)..".mp3",
-			[MAT_SNOW] = "nz_moo/zombies/spawn/snow/pfx_zm_spawn_snow_0"..math.random(0,1)..".mp3",
-			[MAT_SLOSH] = "nz_moo/zombies/spawn/mud/pfx_zm_spawn_mud_00.mp3",
-			[0] = "nz_moo/zombies/spawn/default/pfx_zm_spawn_default_00.mp3",
-		}
-		SpawnMatSound[MAT_GRASS] = SpawnMatSound[MAT_DIRT]
-		SpawnMatSound[MAT_SAND] = SpawnMatSound[MAT_DIRT]
-
-		local norm = (self:GetPos()):GetNormalized()
-		local tr = util.QuickTrace(self:GetPos(), norm*10, self)
-
-		self:SolidMaskDuringEvent(MASK_PLAYERSOLID)
-
-		--ParticleEffect("bo3_zombie_spawn",self:GetPos()+Vector(0,0,1),self:GetAngles(),self)
-		--self:EmitSound("nz/zombies/spawn/zm_spawn_dirt"..math.random(1,2)..".wav",80,math.random(95,105))
-	
-		self:SetSpecialAnimation(true)
-		self:SetIsBusy(true)
-		local seq = self:SelectSpawnSequence()
+    self:SetSubMaterial(2, "vgui/black.vtf")
+    --self:SetModel("models/nzr/ascension_zombies.mdl")
+    local seq = self.EmergeSequences[self:GetEmergeSequenceIndex()]
+    local _, dur = self:LookupSequence(seq)
 
 
-		local navtypes = {
-			[NAV_MESH_OBSTACLE_TOP] = true,
-			[NAV_MESH_DONT_HIDE] = true,
-		}
+    self:MakeDust()
+    
+    -- play emerge animation on spawn
+    -- if we have a coroutine else just spawn the zombie without emerging for now.
+    if coroutine.running() then
+        self:PlaySequenceAndWait(seq)
+    end
 
+    self:SetLastActive(CurTime())
+end
 
-		if IsValid(nav) and nav:HasAttributes(NAV_MESH_OBSTACLE_TOP) then
-			seq = self.UndercroftSequences[math.random(#self.UndercroftSequences)]
-		elseif IsValid(nav) and nav:HasAttributes(NAV_MESH_DONT_HIDE) then
-			seq = "nz_moo_wall_emerge_quick"
-		else
-			if tr.Hit then
-				local finalsound = SpawnMatSound[tr.MatType] or SpawnMatSound[0]
-				self:EmitSound(finalsound)
-			end
-			ParticleEffect("bo3_zombie_spawn",self:GetPos()+Vector(0,0,1),self:GetAngles(),self)
-			self:EmitSound("nz/zombies/spawn/zm_spawn_dirt"..math.random(1,2)..".wav",80,math.random(95,105))
-		end
-		if seq then
-			if IsValid(nav) and (nav:HasAttributes(NAV_MESH_OBSTACLE_TOP) or nav:HasAttributes(NAV_MESH_DONT_HIDE)) then
-				self:PlaySequenceAndMove(seq, {gravity = false})
-			else
-				self:PlaySequenceAndMove(seq, {gravity = true})
-			end
-			self:SetSpecialAnimation(false)
-			self:SetIsBusy(false)
-			self:CollideWhenPossible()
-		end
-	end
+function ENT:OnZombieDeath(dmgInfo)
+
+    if dmgInfo:GetDamageType() == DMG_SHOCK then
+        self:SetRunSpeed(0)
+        self.loco:SetVelocity(Vector(0,0,0))
+        self:Stop()
+        local seq, dur = self:LookupSequence(self.ElectrocutionSequences[math.random(#self.ElectrocutionSequences)])
+        self:ResetSequence(seq)
+        self:SetCycle(0)
+        self:SetCollisionGroup(COLLISION_GROUP_DEBRIS)
+        -- Emit electrocution scream here when added
+        timer.Simple(dur, function()
+            if IsValid(self) then
+            --  self:BecomeRagdoll(dmgInfo) -- It's causing crashes!
+                self:Kill()
+            end
+        end)
+    else
+        self:EmitSound( self.DeathSounds[ math.random( #self.DeathSounds ) ], 100)
+        self:BecomeRagdoll(dmgInfo)
+    end
+
 end
